drop to lowest level to refactor the main menu. Will now draw sidebar as static on main window and keep a stack to the right to swap out. Also shifting code to more separate files. This means more imports but it should help keep things clean and readable. and maintainable.
